[feature/6-refactoring-project] 프로젝트 상세 수정
